ISSUE #3584 - create type of jobs
